Add TLS echo server
The TLS echo server create a TCP listener and a TLS listener based on
the former. It configures the TLS layer to use P-256 elliptic curve
based key exchange, TLS 1.2 minimum, and to prefer server-side ciphers.
It also loads a certificate and a private key from specified files and
supports on-demand shutdown.
Takeaways:
- client interaction logic is implemented in the same way, agnostic to
  whether or not the connection is plain TCP or TLS over TCP;
- certificates + private keys are represented as a single
  tls.Certificate entities, although stored in separate files;
- it's common practice to provide NewMyStruct functions for creating
  instances