Add utility to generate certificate and private key
The utility creates a certificate template for key exchange, signing
data, and certificate signing, that can be used both for client and
server authentication. It generates a private-public key pair, using
P-256 elliptic curve, and a self-signed certificate in DER-format, based
on the keys. It then saves the certificate and the private key in
PEM-format in the specified files.

The certificate is generated to support specified host names and
IP-addresses. It is issued by my own non-existing organization, so,
obviously, can't be used outside of my testing environment.

Takeaways:
- gained some understanding of how certificates and keys are generated
  and how they are related to each other;
- learnt about formats, commonly used to represent certificates and keys
  in binary and in plain text